+++ src/Polynomials/LegendrePolynomialBases.jl
@@ -1,52 +1,60 @@
-struct JacobiPolynomial <: Field end
+"""
+    LegendrePType{K} <: PolynomialType{K}
 
-struct JacobiPolynomialBasis{D,T} <: AbstractVector{JacobiPolynomial}
+Type representing Legendre polynomials of order up to `K`
+"""
+struct LegendrePType{K} <: PolynomialType{K} end
+isHierarchical(::LegendrePType) = true
+
+struct LegendrePolynomial <: Field end
+
+struct LegendrePolynomialBasis{D,T} <: AbstractVector{LegendrePolynomial}
   orders::NTuple{D,Int}
   terms::Vector{CartesianIndex{D}}
-  function JacobiPolynomialBasis{D}(
+  function LegendrePolynomialBasis{D}(
     ::Type{T}, orders::NTuple{D,Int}, terms::Vector{CartesianIndex{D}}) where {D,T}
     new{D,T}(orders,terms)
   end
 end
 
-@inline Base.size(a::JacobiPolynomialBasis{D,T}) where {D,T} = (length(a.terms)*num_indep_components(T),)
-@inline Base.getindex(a::JacobiPolynomialBasis,i::Integer) = JacobiPolynomial()
-@inline Base.IndexStyle(::JacobiPolynomialBasis) = IndexLinear()
+@inline Base.size(a::LegendrePolynomialBasis{D,T}) where {D,T} = (length(a.terms)*num_indep_components(T),)
+@inline Base.getindex(a::LegendrePolynomialBasis,i::Integer) = LegendrePolynomial()
+@inline Base.IndexStyle(::LegendrePolynomialBasis) = IndexLinear()
 
-function JacobiPolynomialBasis{D}(
+function LegendrePolynomialBasis{D}(
   ::Type{T}, orders::NTuple{D,Int}, filter::Function=_q_filter) where {D,T}
 
   terms = _define_terms(filter, orders)
-  JacobiPolynomialBasis{D}(T,orders,terms)
+  LegendrePolynomialBasis{D}(T,orders,terms)
 end
 
-function JacobiPolynomialBasis{D}(
+function LegendrePolynomialBasis{D}(
   ::Type{T}, order::Int, filter::Function=_q_filter) where {D,T}
 
   orders = tfill(order,Val{D}())
-  JacobiPolynomialBasis{D}(T,orders,filter)
+  LegendrePolynomialBasis{D}(T,orders,filter)
 end
 
 # API
 
-function get_exponents(b::JacobiPolynomialBasis)
+function get_exponents(b::LegendrePolynomialBasis)
   indexbase = 1
   [Tuple(t) .- indexbase for t in b.terms]
 end
 
-function get_order(b::JacobiPolynomialBasis)
+function get_order(b::LegendrePolynomialBasis)
   maximum(b.orders)
 end
 
-function get_orders(b::JacobiPolynomialBasis)
+function get_orders(b::LegendrePolynomialBasis)
   b.orders
 end
 
-return_type(::JacobiPolynomialBasis{D,T}) where {D,T} = T
+return_type(::LegendrePolynomialBasis{D,T}) where {D,T} = T
 
 # Field implementation
 
-function return_cache(f::JacobiPolynomialBasis{D,T},x::AbstractVector{<:Point}) where {D,T}
+function return_cache(f::LegendrePolynomialBasis{D,T},x::AbstractVector{<:Point}) where {D,T}
   @check D == length(eltype(x)) "Incorrect number of point components"
   np = length(x)
   ndof = length(f)
@@ -57,7 +65,7 @@ function return_cache(f::JacobiPolynomialBasis{D,T},x::AbstractVector{<:Point}) 
   (r, v, c)
 end
 
-function evaluate!(cache,f::JacobiPolynomialBasis{D,T},x::AbstractVector{<:Point}) where {D,T}
+function evaluate!(cache,f::LegendrePolynomialBasis{D,T},x::AbstractVector{<:Point}) where {D,T}
   r, v, c = cache
   np = length(x)
   ndof = length(f)
@@ -67,7 +75,7 @@ function evaluate!(cache,f::JacobiPolynomialBasis{D,T},x::AbstractVector{<:Point
   setsize!(c,(D,n))
   for i in 1:np
     @inbounds xi = x[i]
-    _evaluate_nd_jp!(v,xi,f.orders,f.terms,c)
+    _evaluate_nd_leg!(v,xi,f.orders,f.terms,c)
     for j in 1:ndof
       @inbounds r[i,j] = v[j]
     end
@@ -76,7 +84,7 @@ function evaluate!(cache,f::JacobiPolynomialBasis{D,T},x::AbstractVector{<:Point
 end
 
 function return_cache(
-  fg::FieldGradientArray{1,JacobiPolynomialBasis{D,V}},
+  fg::FieldGradientArray{1,LegendrePolynomialBasis{D,V}},
   x::AbstractVector{<:Point}) where {D,V}
 
   f = fg.fa
@@ -95,7 +103,7 @@ end
 
 function evaluate!(
   cache,
-  fg::FieldGradientArray{1,JacobiPolynomialBasis{D,T}},
+  fg::FieldGradientArray{1,LegendrePolynomialBasis{D,T}},
   x::AbstractVector{<:Point}) where {D,T}
 
   f = fg.fa
@@ -109,7 +117,7 @@ function evaluate!(
   setsize!(g,(D,n))
   for i in 1:np
     @inbounds xi = x[i]
-    _gradient_nd_jp!(v,xi,f.orders,f.terms,c,g,T)
+    _gradient_nd_leg!(v,xi,f.orders,f.terms,c,g,T)
     for j in 1:ndof
       @inbounds r[i,j] = v[j]
     end
@@ -118,7 +126,7 @@ function evaluate!(
 end
 
 function return_cache(
-  fg::FieldGradientArray{2,JacobiPolynomialBasis{D,V}},
+  fg::FieldGradientArray{2,LegendrePolynomialBasis{D,V}},
   x::AbstractVector{<:Point}) where {D,V}
 
   f = fg.fa
@@ -138,7 +146,7 @@ end
 
 function evaluate!(
   cache,
-  fg::FieldGradientArray{2,JacobiPolynomialBasis{D,T}},
+  fg::FieldGradientArray{2,LegendrePolynomialBasis{D,T}},
   x::AbstractVector{<:Point}) where {D,T}
 
   f = fg.fa
@@ -153,7 +161,7 @@ function evaluate!(
   setsize!(h,(D,n))
   for i in 1:np
     @inbounds xi = x[i]
-    _hessian_nd_jp!(v,xi,f.orders,f.terms,c,g,h,T)
+    _hessian_nd_leg!(v,xi,f.orders,f.terms,c,g,h,T)
     for j in 1:ndof
       @inbounds r[i,j] = v[j]
     end
@@ -163,7 +171,7 @@ end
 
 # Optimizing evaluation at a single point
 
-function return_cache(f::AbstractVector{JacobiPolynomial},x::Point)
+function return_cache(f::AbstractVector{LegendrePolynomial},x::Point)
   xs = [x]
   cf = return_cache(f,xs)
   v = evaluate!(cf,f,xs)
@@ -171,7 +179,7 @@ function return_cache(f::AbstractVector{JacobiPolynomial},x::Point)
   r, cf, xs
 end
 
-function evaluate!(cache,f::AbstractVector{JacobiPolynomial},x::Point)
+function evaluate!(cache,f::AbstractVector{LegendrePolynomial},x::Point)
   r, cf, xs = cache
   xs[1] = x
   v = evaluate!(cf,f,xs)
@@ -183,7 +191,7 @@ function evaluate!(cache,f::AbstractVector{JacobiPolynomial},x::Point)
 end
 
 function return_cache(
-  f::FieldGradientArray{N,<:AbstractVector{JacobiPolynomial}}, x::Point) where {N}
+  f::FieldGradientArray{N,<:AbstractVector{LegendrePolynomial}}, x::Point) where {N}
   xs = [x]
   cf = return_cache(f,xs)
   v = evaluate!(cf,f,xs)
@@ -192,7 +200,7 @@ function return_cache(
 end
 
 function evaluate!(
-  cache, f::FieldGradientArray{N,<:AbstractVector{JacobiPolynomial}}, x::Point) where {N}
+  cache, f::FieldGradientArray{N,<:AbstractVector{LegendrePolynomial}}, x::Point) where {N}
   r, cf, xs = cache
   xs[1] = x
   v = evaluate!(cf,f,xs)
@@ -205,44 +213,53 @@ end
 
 # Helpers
 
-function _evaluate_1d_jp!(v::AbstractMatrix{T},x,order,d) where T
-  n = order + 1
-  z = one(T)
-  @inbounds v[d,1] = z
+function _evaluate_1d!(
+  ::Type{LegendrePType{K}}, v::AbstractMatrix{T},x,d) where {K,T<:Number}
+
+  n = K + 1
+  o = one(T)
+  @inbounds v[d,1] = o
   if n > 1
     ξ = ( 2*x[d] - 1 )
     for i in 2:n
+      # The sqrt(2i-1) factor normalizes the basis polynomial for L2 scalar product on ξ∈[0,1], indeed:
+      # ∫[0,1] Pn(2ξ-1)^2 dξ = 1/2 ∫[-1,1] Pn(t)^2 dt = 1/(2n+1)
+      # C.f. Eq. (1.25) in Section 1.1.5 in Ern & Guermond book (2013).
       @inbounds v[d,i] = sqrt(2*i-1)*jacobi(ξ,i-1,0,0)
     end
   end
 end
 
-function _gradient_1d_jp!(v::AbstractMatrix{T},x,order,d) where T
-  n = order + 1
+function _gradient_1d!(
+  ::Type{LegendrePType{K}}, g::AbstractMatrix{T},x,d) where {K,T<:Number}
+
+  n = K + 1
   z = zero(T)
-  @inbounds v[d,1] = z
+  @inbounds g[d,1] = z
   if n > 1
     ξ = ( 2*x[d] - 1 )
     for i in 2:n
-      @inbounds v[d,i] = sqrt(2*i-1)*i*jacobi(ξ,i-2,1,1)
+      @inbounds g[d,i] = sqrt(2*i-1)*i*jacobi(ξ,i-2,1,1)
     end
   end
 end
 
-function _hessian_1d_jp!(v::AbstractMatrix{T},x,order,d) where T
-  n = order + 1
+function _hessian_1d!(
+  ::Type{LegendrePType{K}}, h::AbstractMatrix{T},x,d) where {K,T<:Number}
+
+  n = K + 1
   z = zero(T)
-  @inbounds v[d,1] = z
+  @inbounds h[d,1] = z
   if n > 1
-    @inbounds v[d,2] = z
+    @inbounds h[d,2] = z
     ξ = ( 2*x[d] - 1 )
     for i in 3:n
-      @inbounds v[d,i] = sqrt(2*i-1)*(i*(i+1)/2)*jacobi(ξ,i-3,2,2)
+      @inbounds h[d,i] = sqrt(2*i-1)*(i*(i+1)/2)*jacobi(ξ,i-3,2,2)
     end
   end
 end
 
-function _evaluate_nd_jp!(
+function _evaluate_nd_leg!(
   v::AbstractVector{V},
   x,
   orders,
@@ -251,7 +268,7 @@ function _evaluate_nd_jp!(
 
   dim = D
   for d in 1:dim
-    _evaluate_1d_jp!(c,x,orders[d],d)
+    _evaluate_1d!(LegendrePType{orders[d]},c,x,d)
   end
 
   o = one(T)
@@ -270,7 +287,7 @@ function _evaluate_nd_jp!(
 
 end
 
-function _gradient_nd_jp!(
+function _gradient_nd_leg!(
   v::AbstractVector{G},
   x,
   orders,
@@ -281,8 +298,7 @@ function _gradient_nd_jp!(
 
   dim = D
   for d in 1:dim
-    _evaluate_1d_jp!(c,x,orders[d],d)
-    _gradient_1d_jp!(g,x,orders[d],d)
+    _derivatives_1d!(LegendrePType{orders[d]},(c,g),x,d)
   end
 
   z = zero(Mutable(VectorValue{D,T}))
@@ -311,7 +327,7 @@ function _gradient_nd_jp!(
 
 end
 
-function _hessian_nd_jp!(
+function _hessian_nd_leg!(
   v::AbstractVector{G},
   x,
   orders,
@@ -323,9 +339,7 @@ function _hessian_nd_jp!(
 
   dim = D
   for d in 1:dim
-    _evaluate_1d_jp!(c,x,orders[d],d)
-    _gradient_1d_jp!(g,x,orders[d],d)
-    _hessian_1d_jp!(h,x,orders[d],d)
+    _derivatives_1d!(LegendrePType{orders[d]},(c,g,h),x,d)
   end
 
   z = zero(Mutable(TensorValue{D,D,T}))