+++ src/Polynomials/QGradMonomialBases.jl
@@ -150,7 +150,7 @@ function _evaluate_nd_qgrad!(
 
   dim = D
   for d in 1:dim
-    _evaluate_1d!(c,x,order,d)
+    _evaluate_1d!(MonomialPType{order},c,x,d)
   end
 
   o = one(T)
@@ -194,8 +194,7 @@ function _gradient_nd_qgrad!(
 
   dim = D
   for d in 1:dim
-    _evaluate_1d!(c,x,order,d)
-    _gradient_1d!(g,x,order,d)
+    _derivatives_1d!(MonomialPType{order},(c,g),x,d)
   end
 
   z = zero(Mutable(V))