+++ src/ReferenceFEs/deprecated/RaviartThomasRefFEs.jl
@@ -16,7 +16,7 @@ is in the Q space of degree `order`.
 function RaviartThomasRefFE(
   ::Type{et},p::Polytope,order::Integer;basis_type=:monomial,phi=GenericField(identity)
 ) where et
-  @assert basis_type ∈ (:monomial, :jacobi, :chebyshev)
+  @assert basis_type ∈ (:monomial, :legendre, :chebyshev)
 
   D = num_dims(p)
 
@@ -24,10 +24,10 @@ function RaviartThomasRefFE(
     prebasis = QCurlGradMonomialBasis{D}(et,order)
   elseif is_simplex(p) && basis_type == :monomial
     prebasis = PCurlGradMonomialBasis{D}(et,order)
-  elseif is_n_cube(p) && basis_type == :jacobi
-    prebasis = QCurlGradJacobiPolynomialBasis{D}(et,order)
-  elseif is_simplex(p) && basis_type == :jacobi
-    prebasis = PCurlGradJacobiPolynomialBasis{D}(et,order)
+  elseif is_n_cube(p) && basis_type == :legendre
+    prebasis = QCurlGradLegendrePolynomialBasis{D}(et,order)
+  elseif is_simplex(p) && basis_type == :legendre
+    prebasis = PCurlGradLegendrePolynomialBasis{D}(et,order)
   elseif is_n_cube(p) && basis_type == :chebyshev
     prebasis = QCurlGradChebyshevPolynomialBasis{D}(et,order)
   else
@@ -196,7 +196,7 @@ function _RT_face_values(p,et,order,phi)
   # Moments (fmoments)
   # The RT prebasis is expressed in terms of shape function
   #fshfs = MonomialBasis(et,fp,order)
-  fshfs = JacobiBasis(et,fp,order)
+  fshfs = LegendreBasis(et,fp,order)
   #fshfs = ChebyshevBasis(et,fp,order)
   #fshfs = get_shapefuns(LagrangianRefFE(et,fp,order))
 
@@ -206,17 +206,17 @@ function _RT_face_values(p,et,order,phi)
   return fcips, fmoments
 end
 
-function JacobiBasis(::Type{T},p::Polytope,orders) where T
-  compute_jacobi_basis(T,p,orders)
+function LegendreBasis(::Type{T},p::Polytope,orders) where T
+  compute_legendre_basis(T,p,orders)
 end
-function JacobiBasis(::Type{T},p::Polytope{D},order::Int) where {D,T}
+function LegendreBasis(::Type{T},p::Polytope{D},order::Int) where {D,T}
   orders = tfill(order,Val{D}())
-  JacobiBasis(T,p,orders)
+  LegendreBasis(T,p,orders)
 end
-function compute_jacobi_basis(::Type{T},p::ExtrusionPolytope{D},orders) where {D,T}
+function compute_legendre_basis(::Type{T},p::ExtrusionPolytope{D},orders) where {D,T}
   extrusion = Tuple(p.extrusion)
   terms = _monomial_terms(extrusion,orders)
-  JacobiPolynomialBasis{D}(T,orders,terms)
+  LegendrePolynomialBasis{D}(T,orders,terms)
 end
 
 function ChebyshevBasis(::Type{T},p::Polytope,orders) where T
@@ -251,7 +251,7 @@ function _RT_cell_values(p,et,order,phi)
   # Cell moments, i.e., M(C)_{ab} = q_C^a(xgp_C^b) w_C^b ⋅ ()
   if is_n_cube(p)
     #cbasis = QGradMonomialBasis{num_dims(p)}(et,order-1)
-    cbasis = QGradJacobiPolynomialBasis{num_dims(p)}(et,order-1)
+    cbasis = QGradLegendrePolynomialBasis{num_dims(p)}(et,order-1)
     #cbasis = QGradChebyshevPolynomialBasis{num_dims(p)}(et,order-1)
     #cbasis = get_shapefuns(RaviartThomasRefFE(et,p,order-1))
   elseif is_simplex(p)