+++ test/PolynomialsTests/runtests.jl
@@ -12,7 +12,7 @@ using Test
 
 @testset "ModalC0Bases" begin include("ModalC0BasesTests.jl") end
 
-@testset "JacobiPolynomialBases" begin include("JacobiPolynomialBasesTests.jl") end
+@testset "LegendrePolynomialBases" begin include("LegendrePolynomialBasesTests.jl") end
 
 #@testset "ChangeBasis" begin include("ChangeBasisTests.jl") end